requirements.txt
ruamel.yaml
semver
sphinx
sphinx_rtd_theme
